fix for m_lastbinding errors in texture.
m_lastbinding was being used internally as a cache of the last texture
bound to a target, rather than a cached enum for glget.  This was
inconsistent with the indicated usage in the constructor, is impossible,
and was causing gl errors.  Changes to update m_lastbinding whenever
m_target is updated, and added a function to get just the enum target
for glGet.

Todo: formatting (tabs ins. of spaces, not done here because it would
make the diff harder to read)
Check similar conditions in other gl classes.

Still need to fix for non DSA path, optimize api traces, logging define,
not-safe define, etc.